+++ appinfo/routes.php
@@ -116,5 +116,6 @@ return [
 		['name' => 'mobile_api#listTransactionAttachments', 'url' => '/api/mobile/v1/workspaces/{workspaceId}/transactions/{txId}/attachments', 'verb' => 'GET', 'requirements' => ['workspaceId' => '\\d+', 'txId' => '\\d+']],
 		['name' => 'mobile_api#uploadTransactionAttachment', 'url' => '/api/mobile/v1/workspaces/{workspaceId}/transactions/{txId}/attachments', 'verb' => 'POST', 'requirements' => ['workspaceId' => '\\d+', 'txId' => '\\d+']],
 		['name' => 'mobile_api#deleteTransactionAttachment', 'url' => '/api/mobile/v1/workspaces/{workspaceId}/attachments/{attachmentId}', 'verb' => 'DELETE', 'requirements' => ['workspaceId' => '\\d+', 'attachmentId' => '\\d+']],
+		['name' => 'mobile_api#downloadTransactionAttachment', 'url' => '/api/mobile/v1/workspaces/{workspaceId}/attachments/{attachmentId}/download', 'verb' => 'GET', 'requirements' => ['workspaceId' => '\\d+', 'attachmentId' => '\\d+']],
 	],
 ];